examples: combine typescript-custom-packages into a single generate.ts
Address review: merge generate-local.ts and generate-sql-on-fhir.ts into one
generate.ts that runs both input mechanisms (local StructureDefinitions →
fhir-types, remote TGZ → sql-on-fhir-types) in sequence. Update Makefile and READMEs.
